==> 05_BYSY_Biblio.Rmd <==
---
title: "BYSY project Notebook"
output: html_notebook
---  
# 05_Adding bibliographic data to Ori  

***  

In this notebook, we add bibliographic data concerning [RIF1](Hafner et al., 2018), [CTF19](Natsume et al., 2013), [FKH1/FKH2](Knott et al., 2012) to known ARS coming from OriDB and transpose these information to the SY_ARS (see notebook [02_Genome Annotation]()).  

## Importing Ori
```{r message=FALSE, warning=FALSE}
suppressMessages(library(GenomicRanges))
suppressMessages(library(tidyverse))
suppressMessages(library(rtracklayer))
`%+%`<- paste0
bibliodata_path <- "Data_raw/"
ARS_oriDB <- read_tsv("Data_raw/ARSfromOriDB20231204.txt",show_col_types = FALSE) %>%
	mutate(chrRom="chr" %+% as.roman(chr))%>% 
	mutate(chr="chr" %+% chr)
### create a naming for ARS oriDB
ARS_oriDB2 <- ARS_oriDB %>%
	mutate(newname=pmap_chr(., function(name,chrRom,status,...) {
		if (status %in% c("Likely","Dubious"))
			{res=paste(status,chrRom,sep="_")}
			else
			{res=name}
			return(res)
		}))%>%
	group_by(newname) %>%
	mutate(newname2=ifelse((!status %in% c("Likely","Dubious") & name!="ARS302"),name,paste(newname,1:n(),sep="_"))) %>% 
	ungroup %>%
	mutate(name=ifelse(is.na(name),"putative_ARS",name))
```


## RIF1 data
```{r message=FALSE, warning=FALSE}
ARSrif1 <- read_tsv(bibliodata_path %+% "Hafner_RIF1.txt",show_col_types = FALSE) %>%
	mutate(rif1_Hafner=map_chr(rif1_, function(x) case_when(x %in% c("pos")~"rep",x=="neg"~"act",x=="no difference"~"null",T~NA))) %>%
	mutate(rif1RBM=map_chr(rif1_RBM, function(x) case_when(x %in% c("pos")~"act",x=="neg"~"rep",x=="no difference"~"null"))) %>%
	select(-c(4,5,6))
ARS_merge1 <- left_join(ARS_oriDB2,ARSrif1, by=join_by(chrRom==chrom,start,end))
```

## FKH1/FKH2 data
```{r message=FALSE, warning=FALSE}
ARSfkh <- read_tsv(bibliodata_path %+% "Knott_FKH12.txt",skip=1,col_names=c("name","Other_name","Chromosome","start","end","fkh1","fkh2","fkh1fkh2"),show_col_types = FALSE) %>%
	select(-c(6,7,9)) %>% 
	mutate(chrom=paste0("chr",as.roman(Chromosome)))	%>%
	mutate(fkh1fkh2_Knott=map_chr(fkh1fkh2, function(x) 
		case_when(x==1~"rep",x==-1~"act",x==0~"null"))) %>%
	select(-fkh1fkh2)
ARS_merge1.gr <- with(ARS_merge1,GRanges(seqnames=chrRom,range=IRanges(start,end),strand="*",name=newname2))

ARSfkh.gr <- with(ARSfkh,GRanges(seqnames=chrom,range=IRanges(start,end),strand="*",fkh1fkh2_Knott=fkh1fkh2_Knott))

ol1 <- findOverlaps(ARSfkh.gr,ARS_merge1.gr)
ARS_merge2 <- ARS_merge1
ARS_merge2$fkh1fkh2_Knott <- sapply(1:length(ARS_merge1.gr), function(x) {res=ARSfkh.gr[queryHits(ol1)[subjectHits(ol1)==x]]$fkh1fkh2_Knott;if (length(res)==0) {res=NA};return(res)})
```


## CTF19 data
```{r message=FALSE, warning=FALSE}
ARSctf19 <- read_tsv(bibliodata_path %+% "MyCTF19fromWig20230403.txt",show_col_types = FALSE) %>% mutate(start=start+1) %>% mutate(ctf19_Natsume=map_chr(CTF19D, function(x) case_when(x=="p"~"act",T~"null"))) %>% dplyr::rename(chrRom=chrom)

ARS_merge3 <- full_join(ARS_merge2,ARSctf19 %>% select(c(1,2,3,8)),by = join_by(start, end, chrRom))

write_tsv(ARS_merge3,file="Data/ARS_withBiblioData.tsv")
```


## Transpostion to BY_SY Ori
```{r message=FALSE, warning=FALSE}
ARS_BYSY <- readRDS("Data/ARS_BYonSY.rds")
ARS_merge4 <- ARS_merge3 %>% dplyr::rename(chromROM=chrRom)
ARS_merge5 <- full_join(ARS_BYSY,ARS_merge4 %>% select(-name) %>% dplyr::rename(name=newname2,startsc1=start,endsc1=end),by = join_by(name, chromROM)) %>% arrange(chromROM,start)
write_tsv(ARS_merge5,file="Data/ARS_withBiblioData_namesBYSY.tsv")

ARSwCTF19 <- with(ARS_merge5 %>% filter(!is.na(chromSY),ctf19_Natsume=="act"),GRanges(seqnames=chromSY,ranges=IRanges(startnSY,endnSY),name=name))
rtracklayer::export(ARSwCTF19,con="Data/ARS_SYwCTF19.bed")
ARSwRIF1p <- with(ARS_merge5 %>% filter(!is.na(chromSY),rif1_Hafner=="rep"),GRanges(seqnames=chromSY,ranges=IRanges(startnSY,endnSY),name=name))
rtracklayer::export(ARSwRIF1p,con="Data/ARS_SYwRIF1r.bed")
ARSwRIF1n <- with(ARS_merge5 %>% filter(!is.na(chromSY),rif1_Hafner=="act"),GRanges(seqnames=chromSY,ranges=IRanges(startnSY,endnSY),name=name))
rtracklayer::export(ARSwRIF1n,con="Data/ARS_SYwRIF1a.bed")
ARSwFKHa <- with(ARS_merge5 %>% filter(!is.na(chromSY),fkh1fkh2_Knott=="act"),GRanges(seqnames=chromSY,ranges=IRanges(startnSY,endnSY),name=name))
rtracklayer::export(ARSwFKHa,con="Data/ARS_SYwFKHa.bed")
ARSwFKHr <- with(ARS_merge5 %>% filter(!is.na(chromSY),fkh1fkh2_Knott=="rep"),GRanges(seqnames=chromSY,ranges=IRanges(startnSY,endnSY),name=name))
rtracklayer::export(ARSwFKHr,con="Data/ARS_SYwFKHr.bed")
```

*There are duplication in name generated by transposition and split-mapping for Dubious_chrIII_3,Dubious_chrIII_5,Dubious_chrX_2 and ARS1106.5*
